+++ src/utils/s3.ts
@@ -1,12 +1,15 @@
 import AWS from 'aws-sdk';
 
 import Config from '../config';
+import { ReadStream } from 'fs';
 
 const config = Config[Config.env];
 
 export const uploadToS3 = async (
   Key: string,
-  Body: string | Buffer | Blob
+  Body: string | Buffer | Blob | ReadStream,
+  ContentEncoding?: string | undefined,
+  ContentType?: string
 ): Promise<boolean> => {
   try {
     return new Promise(async (resolve) => {
@@ -15,6 +18,8 @@ export const uploadToS3 = async (
         Key,
         ACL: 'public-read',
         Body,
+        ContentEncoding,
+        ContentType,
       };
 
       new AWS.S3().putObject(params, (err, data) => {